fix: suggestions deletion splice is really badly goofed & issue with textbasedchannel

+++ src/server/servers/routes/suggestions.ts
@@ -55,6 +55,8 @@ const suggestions = (auxdibot: Auxdibot, router: Router) => {
          if (!suggestion) return res.status(404).json({ error: 'invalid suggestion' });
          const channel = req.guild.channels.cache.get(server.suggestions_channel);
          if (!channel) return res.status(404).json({ error: 'There is no suggestions channel!' });
+         if (!channel.isTextBased())
+            return res.status(400).json({ error: 'The suggestions channel is not a text channel!' });
          const msg = await channel.messages.fetch(suggestion.messageID).catch(() => undefined);
          if (msg) await msg.delete().catch(() => undefined);
          return deleteSuggestion(auxdibot, req.guild.id, Number(suggestionID))
